backend/requirements.txt
flask
flask-cors
requests
flask-sqlalchemy
pymysql
python-dotenv
certifi
cloudinary
gunicorn
